Move method for copying a `map[string]string`
