+++ src/kvstorage/map.go
@@ -1,8 +1,7 @@
-// Package maputil provides Go maps related utility methods
-package maputil
+package kvstorage
 
 // Copy copies the map contents to the new map
-func Copy(data map[string]string) map[string]string {
+func copyMap(data map[string]string) map[string]string {
 	copied := make(map[string]string, len(data))
 
 	for k, v := range data {
